Bemidji ID Change
Importing the Cities was causing a problem, I ended up having to move
Bemidji to where it already was in the city list to get the City CSV to
import.  So on all exisitng CSV Files City_ID 195 needs to be updated to
24.  There is an error on the import so you can't update it without
fixing the ID so we should be fine.